regularize epitope activities (#89)
This pull request adds regularization to the epitope activities. The regularization is a Pseudo-Huber function for positive activities, and zero for negative activities. This helps keep activities <= 0 for epitopes that are in unimportant. The regularization is done via `reg_activity_weight` parameter passed to `Polyclonal.fit`.

Empirically, this improves fitting speed a bit and seems to otherwise give comparable-ish results. The default weight is 1.

@timcyu, I did not re-run your notebooks on selecting numbers of epitopes etc after this change, but I expect it will have only a modest effect. Nonetheless, this provide a way to now adjust regularization on the other important model parameter. 

Specific changes:

* add equations for normalizing wildtype activity

* implement regularization of epitope activities

* re-run notebooks with regularized activity

* fix epitope harmony test after regularizing weights