generalize `mut_escape_heatmap` (#92)
Add some options that generalize `mut_escape_heatmap` for more plotting uses. This enables it to be used for instance to plot functional scores as well with all the nice interactivity. This isn't really a specific targeted use of the `polyclonal` package, but makes the API more broadly useful.

Specifically:
  * added `epitope_label_suffix` parameter
  * added `diverging_colors` parameter
  * changed `percent_max_cutoff` slider to work on real rather than absolute values and with non-zero minimum values.

This increments version to 0.2.