generalize `PolyclonalCollection` and implement `PolyclonalBootstrap` (#96)
This pull request implements step 1 of #95 by making `PolyclonalCollection` a general class that is subclassed by `PolyclonalBootstrap`. This better generalizes the idea of a collection as described in #95. Before `PolyclonalCollection` was named like a general-purpose base class, but then implemented as a specific-use-case class that could not plausibly be used as a base class.

This will be part of version 1.0 and is backwards incompatible.

@zorian15 @matsen, just flagging as it changes a bit about the bootstrapping interface.