replace `Polyclonal.harmonize_epitopes_with` with `Polyclonal.epitope_harmonized_model`.
The old method for harmonizing epitopes, `Polyclonal.harmonize_epitopes_with`, had several shortcomings. It modified the model in place, and did not guarantee consistent epitope order. It worked with bootstrapping when the models to be harmonized were only created for that purpose, but won't work for things like model averaging.

The new `Polyclonal.epitope_harmonized_model` returns a copy of the original model and guarantees harmonized models in consistent order, and so is suitable for purposes like model averaging.